Move `theme.ts` out of `Layout`
