+++ components/index.ts
@@ -8,6 +8,7 @@ export * from './PaginationSection'
 export * from './SearchBar'
 export * from './ServiceItem'
 export * from './ServicesListItem'
+export * from './ShareLink'
 export * from './SortMenu'
 export * from './TaxonomySection'
-export * from './ShareLink'
+export * from './theme'
